Separate flash_n300_multihost into its own step
There is some bug related to the lifespan of the chip_objs in
connection_data which was causing spi_writes in flash_n300_multihost
to fail, but appear to succeed. We need to look deeper into this, but
for now adding another reset before flash_n300_multihost and
re-detecting the chips works.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>