Address single-host assumption in generate_connection_map
Skip remote chip IDs from ETH FW that weren't originally detected in the
construction of the chip_data dict in generate_connection_map.

This allows systems with cards that are linked up to cards on another
host to be configured in a single-host mesh configuration. Before,
attempting to do this caused a KeyError, as it assumed that all non-zero
remote_infos would be valid keys to chip_data.

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>