Add mesh_v2 layout option and multi-host programming function
mesh_v2 topology is a request from the product team to allow us to use a
different cable setup on Wormhole LoudBox.

This is the same as the regular mesh topology, except:
- ethernet coordinates are swapped between coords (1,0) and (1,1)
- intermesh prorgramming is swapped between coords (2,1) and (2,0)

Using tt-topology -l mesh_v2 flashes the regular mesh topology, then performs
the v2 multi-host configuration, which includes the eth coordinate swap between
(1,0) and (1,1).

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>